+++ src/utils.rs
@@ -1,2 +1,3 @@
+pub mod scrubbing;
 pub mod logging;
 pub mod test;
